requirements.txt
asttokens==2.4.1
cloudpickle==3.0.0
colorama==0.4.6
comm==0.2.2
debugpy==1.8.1
decorator==5.1.1
executing==2.0.1
ipykernel==6.29.4
ipython==8.25.0
jedi==0.19.1
jupyter_client==8.6.2
jupyter_core==5.7.2
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==1.26.4
opencv-python==4.10.0.82
packaging==24.0
parso==0.8.4
platformdirs==4.2.2
prompt_toolkit==3.0.46
psutil==5.9.8
pure-eval==0.2.2
Pygments==2.18.0
python-dateutil==2.9.0.post0
pywin32==306
pyzmq==26.0.3
six==1.16.0
spyder-kernels==2.5.1
stack-data==0.6.3
tornado==6.4
traitlets==5.14.3
wcwidth==0.2.13
